Introduce PageCache backend for Lucene
Note: This is an unsanctioned proposal, if you are an outside
reader please do not expect this to be merged.

Kernel: As discussed. Benchmarks forthcoming, opening PR to discuss
feasibility of this. Suggested end-game, if we decide to do this, would
be to introduce this behind an unsanctioned toggle, pickle in
soak testing for a few months and then evaluate.

Overview & Safety
--

This provides an implementation of the Lucene Directory API that uses
`FileSystemAbstraction` for writes and `PageCache` for reads.

It is modelled after, and is intended to inherit the safety aspects of,
the `MMapDirectory` that ships with Lucene. In fact, it is precisely the
same as `MMapDirectory`, except is uses the `PageCache` as its buffer
backend for index inputs.

Since Lucene writes to immutable files, and since it does not read
from those files before they are fully written, this implementation
has the same safety properties as MMapDirectory: Safety comes from
proper fsyncing of writes, and this implementation shares the exact
same write behavior as MMapDirectory.

Hence: Review of this should, outside of regular crash fuzzing, verify
that the above actually holds true: The write-side implementation should
be equivalent to that of MMapDirectory.

Background
--

Lucene provides multiple storage backends:

- RAMDirectory
    in-memory, used by kernel in Ephemeral mode

- NIOFSDirectory
    persistent, loads fixed chunks of indexes into a buffer
    as they are read; meaning lots of IO. Used by kernel if
    -Dorg.neo4j.kernel.api.impl.index.storage.DirectoryFactory.default_directory_factory=true

- MMapDirectory
    persistent, memory maps whole indexes as they are read,
    used by kernel by default.

RAMDirectory is not persistent. NIOFSDirectory is slow, so MMapDirectory
is used by default.

However, there are issues with MMapDirectory as well:

*1. Swapping*

Because MMapDirectory memory maps entire indexes, any non-trivial
deployment will be forced to swap the Neo4j process. This introduces at
least two issues. 1, there is no way to operate Neo4j in an environment
without swap space and, 2, memory management of Neo4j is expanded to
require operators to consider their virtual memory limits on day one.

Since standard AWS images do not have swap enabled by default, this
poses a substantial UX pain: Most users will experience this by the
Neo4j process simply consuming indefinite amounts of memory, with no
indication that turning on swap would solve the problem.

There may also be problems with this swapping then inducing competition
between hot indexes and runtime data structures in the database itself,
although this isn't entirely clear.

*2. Java MappedByteBuffer*

As we are acutely aware, the Java memory mapping APIs have some
undesireable characteristics. Lucene tries to alleviate this by
using the Cleaner API.
While this somewhat reduces GC pain, the BBs objects are still stuck
on ReferenceQueues by the Cleaner API, increasing GC work during
safepoints. It also means we're implicitly calling unsafe.freeMemory(address),
and memory mapping, over and over as index readers are opened and closed.

Rationale
---

This implementation thus provides two benefits:

1) It allows Neo4j to run in environments with no or limited Swap
2) It gives Neo4j predictable and configurable memory use
3) It removes substantial numbers of `mmap/dealloc` from hot paths